'separate partial_dense_KPConv'
