+++ models/KPConv/modules.py
@@ -31,12 +31,12 @@ from torch_scatter import scatter_max
 # Adaption from https://github.com/humanpose1/KPConvTorch/blob/master/models/layers.py
 from .kernel_utils import load_kernels as create_kernel_points
 from .convolution_ops import KPConv_ops, KPConv_deform_ops
-from .utils import weight_variable
 from .kernels import PointKernel, LightDeformablePointKernel
 from .kernel_utils import kernel_point_optimization_debug
 from models.core_sampling_and_search import RadiusNeighbourFinder, FPSSampler
 from models.core_modules import *
 from models.unet_base import BaseFactory
+from .partial_dense_modules import *
 
 
 class KPConvModels(Enum):
@@ -66,159 +66,6 @@ class KPConvFactory(BaseFactory):
                 return getattr(self.modules_lib, self.module_name_down, None)
 
         raise NotImplementedError
-
-
-####################### BUILT WITH PARTIAL DENSE FORMAT ############################
-
-
-class BaseKPConvPartialDense(BaseConvolutionDown):
-    def __init__(
-        self,
-        ratio=None,
-        radius=None,
-        down_conv_nn=None,
-        kp_points=16,
-        nb_feature=0,
-        is_strided=True,
-        KP_EXTENT=None,
-        DENSITY_PARAMETER=None,
-        *args,
-        **kwargs
-    ):
-        super(BaseKPConvPartialDense, self).__init__(
-            FPSSampler(ratio), RadiusNeighbourFinder(radius, conv_type=kwargs.get("conv_type")), *args, **kwargs
-        )
-
-        self.ratio = ratio
-        self.radius = radius
-        self.is_strided = is_strided
-
-        if len(down_conv_nn) == 2:
-            in_features, out_features = down_conv_nn
-            intermediate_features = None
-
-        elif len(down_conv_nn) == 3:
-            in_features, intermediate_features, out_features = down_conv_nn
-
-        else:
-            raise NotImplementedError
-
-        # KPCONV arguments
-        self.in_features = in_features
-        self.out_features = out_features
-        self.intermediate_features = intermediate_features
-        self.kp_points = kp_points
-
-        # Dataset ~ Model parameters
-        self.KP_EXTENT = KP_EXTENT
-        self.DENSITY_PARAMETER = DENSITY_PARAMETER
-
-        # PARAMTERS IMPORTANT FOR SHADOWING
-        self.shadow_features_fill = 0.0
-        self.shadow_points_fill_ = float(10e6)
-
-
-class KPConvPartialDense(BaseKPConvPartialDense):
-    def __init__(self, *args, **kwargs):
-        super(KPConvPartialDense, self).__init__(*args, **kwargs)
-
-        self._conv = PointKernelPartialDense(
-            self.kp_points,
-            self.in_features,
-            self.out_features,
-            radius=self.radius,
-            is_strided=self.is_strided,
-            KP_EXTENT=self.KP_EXTENT,
-            DENSITY_PARAMETER=self.DENSITY_PARAMETER,
-        )
-        self.activation = kwargs.get("act", nn.LeakyReLU(0.2))
-
-    def conv_partial_dense(self, input, pos, input_neighbour, pos_neighbour, idx_neighbour, idx_sampler):
-        return self._conv(input_neighbour, pos_neighbour, idx_sampler)
-
-
-class ResnetPartialDense(BaseKPConvPartialDense):
-    def __init__(self, *args, **kwargs):
-        super(ResnetPartialDense, self).__init__(*args, **kwargs)
-
-        self._kp_conv0 = PointKernelPartialDense(
-            self.kp_points,
-            self.in_features,
-            self.intermediate_features,
-            radius=self.radius,
-            is_strided=False,
-            KP_EXTENT=self.KP_EXTENT,
-            DENSITY_PARAMETER=self.DENSITY_PARAMETER,
-        )
-
-        self._kp_conv1 = PointKernelPartialDense(
-            self.kp_points,
-            self.intermediate_features,
-            self.out_features,
-            radius=self.radius,
-            is_strided=self.is_strided,
-            KP_EXTENT=self.KP_EXTENT,
-            DENSITY_PARAMETER=self.DENSITY_PARAMETER,
-        )
-
-        if self.out_features != self.intermediate_features:
-            self.shortcut_op = UnaryConv(self.intermediate_features, self.out_features)
-        else:
-            self.shortcut_op = torch.nn.Identity()
-
-    def conv_partial_dense(self, input, pos, input_neighbour, pos_centered_neighbour, idx_neighbour, idx_sampler):
-
-        x = self._kp_conv0(input, idx_neighbour, pos_centered_neighbour, idx_sampler=None)
-        x = self._kp_conv1(x, idx_neighbour, pos_centered_neighbour, idx_sampler=idx_sampler)
-
-        if self.is_strided:
-            input = input_neighbour[idx_sampler].max(1)[0]
-        x = x + self.shortcut_op(input)
-
-        return x
-
-
-class ResnetBottleNeckPartialDense(BaseKPConvPartialDense):
-    def __init__(self, *args, **kwargs):
-        super(ResnetBottleNeckPartialDense, self).__init__(*args, **kwargs)
-
-        self._kp_conv0 = PointKernelPartialDense(
-            self.kp_points,
-            self.intermediate_features,
-            self.intermediate_features,
-            radius=self.radius,
-            is_strided=False,
-            KP_EXTENT=self.KP_EXTENT,
-            DENSITY_PARAMETER=self.DENSITY_PARAMETER,
-        )
-
-        self._kp_conv1 = PointKernelPartialDense(
-            self.kp_points,
-            self.intermediate_features,
-            self.out_features,
-            radius=self.radius,
-            is_strided=self.is_strided,
-            KP_EXTENT=self.KP_EXTENT,
-            DENSITY_PARAMETER=self.DENSITY_PARAMETER,
-        )
-
-        self.uconv_0 = UnaryConv(self.in_features, self.intermediate_features)
-
-        if self.out_features != self.intermediate_features:
-            self.shortcut_op = UnaryConv(self.in_features, self.out_features)
-        else:
-            self.shortcut_op = torch.nn.Identity()
-
-    def conv_partial_dense(self, input, pos, input_neighbour, pos_centered_neighbour, idx_neighbour, idx_sampler):
-
-        x = self.uconv_0(input)
-        x = self._kp_conv0(x, idx_neighbour, pos_centered_neighbour, idx_sampler=None)
-        x = self._kp_conv1(x, idx_neighbour, pos_centered_neighbour, idx_sampler=idx_sampler)
-
-        if self.is_strided:
-            input = input_neighbour[idx_sampler].max(1)[0]
-        x = x + self.shortcut_op(input)
-        return x
 
 
 ####################### BUILT WITH BaseConvolutionDown ############################
@@ -540,27 +387,6 @@ class DeformableKPConvLayer(torch.nn.Module):
         )
         self.sq_distances = torch.nn.Parameter(sq_distances)
         return feat
-
-
-class UnaryConv(torch.nn.Module):
-    def __init__(self, num_inputs, num_outputs):
-        """
-        1x1 convolution on point cloud (we can even call it a mini pointnet)
-        """
-        super(UnaryConv, self).__init__()
-        self.num_inputs = num_inputs
-        self.num_outputs = num_outputs
-
-        self.weight = Parameter(weight_variable([self.num_inputs, self.num_outputs]))
-
-    def forward(self, features):
-        """
-        features(Torch Tensor): size N x d d is the size of inputs
-        """
-        return torch.matmul(features, self.weight)
-
-    def __repr__(self):
-        return "UnaryConv({}, {})".format(self.num_inputs, self.num_outputs)
 
 
 def max_pool(features, pools):
